indexページのリストタイトルcssを修正

+++ app/controllers/list_controller.rb
@@ -34,8 +34,10 @@ class ListController < ApplicationController
   end
 
   def destroy
-    public_ids = [@list.audio.file.public_id]
-    Cloudinary::Api.delete_resources(public_ids)
+    if @list.audio.present?
+      public_ids = [@list.audio.file.public_id]
+      Cloudinary::Api.delete_resources(public_ids)
+    end
     @list.destroy
     flash[:danger] = "リストを削除しました"
     redirect_to list_index_path
